lib/ukfile: API: Change file offset to size_t
This change makes the ukfile API natively use size_t as type for
file offsets. The previous use of the signed off_t was unwieldy, as
negative file offsets had no meaning or use in the internal API.
Best to correct this mistake earlier than later.
This makes posix-fdio syscalls the only entry points with signed file
offset args, allowing validation to be isolated.

+++ lib/posix-eventfd/eventfd.c
@@ -36,7 +36,7 @@ struct evfd_alloc {
 
 static ssize_t evfd_read(const struct uk_file *f,
 			 const struct iovec *iov, int iovcnt,
-			 off_t off, long flags __unused)
+			 size_t off, long flags __unused)
 {
 	int semaphore;
 	evfd_node n;
@@ -76,7 +76,7 @@ static ssize_t evfd_read(const struct uk_file *f,
 
 static ssize_t evfd_write(const struct uk_file *f,
 			  const struct iovec *iov, int iovcnt,
-			  off_t off, long flags __unused)
+			  size_t off, long flags __unused)
 {
 	uint64_t add;
 	evfd_node n;
