+++ lib/posix-socket/socket.c
@@ -94,7 +94,7 @@ static struct uk_ofile *socketfd_get(int fd)
 static ssize_t
 socket_read(const struct uk_file *sock,
 	    const struct iovec *iov, int iovcnt,
-	    off_t off, long flags __unused)
+	    size_t off, long flags __unused)
 {
 	ssize_t ret;
 	struct posix_socket_driver *d;
@@ -126,7 +126,7 @@ socket_read(const struct uk_file *sock,
 static ssize_t
 socket_write(const struct uk_file *sock,
 	     const struct iovec *iov, int iovcnt,
-	     off_t off, long flags __unused)
+	     size_t off, long flags __unused)
 {
 	ssize_t ret;
 	struct posix_socket_driver *d;
